+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-## 0.0.1
+## 1.0.1
 
-* TODO: Describe initial release.
+* Added `FormatusViewer` to display formatted text (see example about usage)
+* Fixed a bug when appending a new paragraph and then entering text
+
+## 1.0.0
+
+* Initial release